+++ virtool/samples/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, Enum, Integer, String
+from sqlalchemy import BigInteger, Column, DateTime, Enum, Integer, String
 from sqlalchemy.sql.schema import ForeignKey, UniqueConstraint
 
 from virtool.pg.base import Base
@@ -33,7 +33,7 @@ class SampleArtifact(Base):
     sample = Column(String, nullable=False)
     name = Column(String, nullable=False)
     name_on_disk = Column(String)
-    size = Column(Integer)
+    size = Column(BigInteger)
     type = Column(Enum(ArtifactType), nullable=False)
     uploaded_at = Column(DateTime)
 
@@ -51,6 +51,6 @@ class SampleReads(Base):
     sample = Column(String, nullable=False)
     name = Column(String(length=13), nullable=False)
     name_on_disk = Column(String, nullable=False)
-    size = Column(Integer)
+    size = Column(BigInteger)
     upload = Column(Integer, ForeignKey("uploads.id"))
     uploaded_at = Column(DateTime)
